+++ src/pages/help-me/register/index.tsx
@@ -13,7 +13,7 @@ import styles from "@/pages/help-me/register/HelpMeRegister.module.scss";
 import classNames from "classnames/bind";
 import CustomDatePicker from "@/components/common/DatePicker/DatePicker";
 import { Controller, useForm } from "react-hook-form";
-import RootLayout from "@/components/RootLayout/RootLayout";
+import RootLayout from "@/components/common/RootLayout/RootLayout";
 import { ReactElement } from "react";
 
 const cn = classNames.bind(styles);
